refactor: Remove deprecated chat types and update message handling
