--- types/chat-extra.ts
@@ -1,6 +1,0 @@
-import { Message as OriginalMessage } from "./chat";
-
-export interface Message extends OriginalMessage {
-  id: string;
-  model?: string;
-}
